pinecone-chatbot/requirements.txt
# LangChain packages - Latest stable versions (v1.0+) for Python 3.12
langchain>=1.0.0
langchain-core>=1.0.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-classic>=0.1.0
langchain-pinecone>=0.1.0

# Vector Database
pinecone>=5.0.0

# AI/ML
openai>=1.54.0
numpy>=1.26.0

# Web Framework (for future FastAPI)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0

# Utilities
python-dotenv>=1.0.1
tqdm>=4.66.0

# PDF Processing
pypdf>=5.0.0
langchain-text-splitters>=1.0.0
